requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
python-dotenv==1.0.1
pydantic==2.9.2
chromadb==0.5.5
pypdf==4.3.1
requests==2.32.3
prometheus-fastapi-instrumentator==7.0.0
structlog==24.4.0
langchain-text-splitters==0.2.4
python-multipart==0.0.9